w3c-reports/12092014/efl/runned-tests/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc-expected.txt


PASS audio.currentSrc initial value 
PASS audio.currentSrc after setting src attribute "" 
PASS audio.currentSrc after adding source element with src attribute "" 
PASS audio.currentSrc after setting src attribute "." 
PASS audio.currentSrc after adding source element with src attribute "." 
FAIL audio.currentSrc after setting src attribute " " assert_equals: expected "http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:21:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL audio.currentSrc after adding source element with src attribute " " assert_equals: expected "http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:37:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS audio.currentSrc after setting src attribute "data:," 
FAIL audio.currentSrc after adding source element with src attribute "data:," assert_equals: expected "data:," but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:37:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS video.currentSrc initial value 
PASS video.currentSrc after setting src attribute "" 
PASS video.currentSrc after adding source element with src attribute "" 
PASS video.currentSrc after setting src attribute "." 
PASS video.currentSrc after adding source element with src attribute "." 
FAIL video.currentSrc after setting src attribute " " assert_equals: expected "http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:21:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL video.currentSrc after adding source element with src attribute " " assert_equals: expected "http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:37:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS video.currentSrc after setting src attribute "data:," 
FAIL video.currentSrc after adding source element with src attribute "data:," assert_equals: expected "data:," but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc.html:37:24
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
